requirements.txt
fastai==1.0.61
numpy==1.18.5
pandas==1.0.5
Pillow==7.1.2
streamlit==0.64.0
torch==1.5.1
torchvision==0.6.1